requirements.txt
arrow==0.10.0
click==6.7
Flask==0.12.2
google-api-python-client==1.6.4
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2client==2.2.0
pkg-resources==0.0.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
python-dateutil==2.6.1
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.12.2
